Move Drawer to another component.
