+++ src/components/StrategyPage/StrategyPage.tsx
@@ -1,27 +1,19 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { get } from 'lodash';
-import { Drawer } from 'antd';
-import { bindActionCreators, Dispatch } from 'redux';
-import { connect } from 'react-redux';
-import { isFunction } from 'lodash';
 import StrategyHeader from './StrategyHeader';
 import StrategyContainer from './StrategyContainer';
 import { StrategyTypes } from '../../enums/strategies';
 import { StrategyPageWrapper } from './styled';
-import { ClientActions, StrategyEntry, CloseDrawerAction } from '../../reducers/client';
-import { RootState, StandardAction } from '../../reducers/reducerTypes';
+import { StrategyEntry } from '../../reducers/client';
+import DrawerContainer from './Drawer/DrawerContainer';
 
 interface StrategyPageProps {
   clientId: number;
-  drawerOpen: boolean;
-  drawerTitle: string;
 
   pageData: StrategyEntry;
-  closeDrawer?: (title: string) => CloseDrawerAction;
 }
 
 const StrategyPage = (props: StrategyPageProps) => {
-  const { drawerOpen, drawerTitle, closeDrawer } = props;
   const { pageData } = props;
   const superannuation = get(pageData, 'superannuation');
   const pension = get(pageData, 'pension');
@@ -30,11 +22,6 @@ const StrategyPage = (props: StrategyPageProps) => {
   const centrelink = get(pageData, 'centrelink');
   const insurance = get(pageData, 'insurance');
   const estatePlanning = get(pageData, 'estatePlanning');
-  const onCloseDrawer = () => {
-    if (isFunction(closeDrawer)) {
-      closeDrawer('');
-    }
-  };
 
   return (
     <StrategyPageWrapper>
@@ -74,29 +61,9 @@ const StrategyPage = (props: StrategyPageProps) => {
           strategies={estatePlanning.strategies}
         />
       )}
-      <Drawer title={drawerTitle} width={720} onClose={onCloseDrawer} visible={drawerOpen}>
-        <p>Some contents...</p>
-        <p>Some contents...</p>
-        <p>Some contents...</p>
-      </Drawer>
+      <DrawerContainer />
     </StrategyPageWrapper>
   );
 };
 
-const mapStateToProps = (state: RootState) => ({
-  drawerOpen: state.client.get('drawerOpen'),
-  drawerTitle: state.client.get('drawerTitle'),
-});
-
-const mapDispatchToProps = (dispatch: Dispatch<StandardAction<any>>) =>
-  bindActionCreators(
-    {
-      closeDrawer: ClientActions.closeDrawer,
-    },
-    dispatch,
-  );
-
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps,
-)(StrategyPage);
+export default StrategyPage;
